% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/healthyControPresenceChecker-package.r
\docType{package}
\name{healthyControlsChecker}
\alias{healthyControlsChecker}
\title{healthyControlsChecker: A package for verifying the presence of healthy controls in a GEO dataset}
\description{
The healthyControlsChecker package provides one relevant function: healthyControlsCheck()
}
\section{healthyControlsCheck() function}{
  The healthyControlsCheck() function readsin the GEO code of a dataset, and returns true if there's at least a feature containing the healthy controls.
}

