% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/healthyControPresenceChecker.r
\name{healthyControlsCheck}
\alias{healthyControlsCheck}
\title{Function that reads in the GEO code of a dataset, and returns true if there's at least a feature containing the healthy controls.}
\usage{
healthyControlsCheck(datasetGeoCode, verbose = FALSE)
}
\arguments{
\item{datasetGeoCode}{the GEO code of a dataset.}

\item{verbose}{a boolean flag stating if helping messages should be printed or not}
}
\value{
a boolean value
}
\description{
Function that reads in the GEO code of a dataset, and returns true if there's at least a feature containing the healthy controls.
}
\examples{
healthyControlsCheckOutcome <- healthyControlsCheck("GSE3268", FALSE)
}
